Cleanup validation usage
