+++ src/lib/create-remote.ts
@@ -1,11 +1,11 @@
 import {resolve} from './utils';
-import {validateMetadataKey} from './validate';
-const ValidationError = require('loopback').ValidationError;
+import {validateArgs} from './validate';
 
 export function createRemoteMethod(RemoteClass: any, RemoteMethod: any, props: any) {
   let {selector, meta, providers} = props;
   if (meta.isStatic) {
     RemoteClass[selector] = async function(...args: any[]) {
+      await validateArgs(args, RemoteMethod);
       let instance = this;
       let proms = await resolve(instance, providers);
       let isNull = proms.indexOf(null);
@@ -17,6 +17,7 @@ export function createRemoteMethod(RemoteClass: any, RemoteMethod: any, props: a
     };
   } else {
     RemoteClass.prototype[selector] = async function(...args: any[]) {
+      await validateArgs(args, RemoteMethod);
       let instance = this;
       let proms = await resolve(instance, providers);
       let isNull = proms.indexOf(null);
@@ -24,23 +25,6 @@ export function createRemoteMethod(RemoteClass: any, RemoteMethod: any, props: a
         throw Error(
             `cannot instantiate ${RemoteMethod.name} provider at position: ${isNull} is null`);
       }
-
-      function getValid(mdl: any) {
-        return new Promise((res: Function, rej: Function) => {
-          mdl.isValid((isValid: boolean) => {
-            if (isValid) {
-              return res();
-            } else {
-              return rej(new ValidationError(mdl));
-            }
-          });
-        });
-      }
-
-      let validatesMeta =
-          Reflect.getOwnMetadata(validateMetadataKey, RemoteMethod, 'onRemote') || [];
-      let validations = validatesMeta.map((idx: number) => getValid(args[idx]));
-      await Promise.all(validations);
 
       return new RemoteMethod(...proms).onRemote(...args);
     };
