+++ src/lib/validate.ts
@@ -1,6 +1,11 @@
 import 'reflect-metadata';
 
+declare var global: any /** TODO #9100 */;
+const Reflect = global['Reflect'];
+
 export const validateMetadataKey = Symbol('validates');
+
+const ValidationError = require('loopback').ValidationError;
 
 export function Validate(target: Object, propertyKey: string|symbol, parameterIndex: number) {
   let existingRequiredParameters: number[] =
@@ -8,4 +13,22 @@ export function Validate(target: Object, propertyKey: string|symbol, parameterIn
   existingRequiredParameters.push(parameterIndex);
   Reflect.defineMetadata(
       validateMetadataKey, existingRequiredParameters, target.constructor, propertyKey);
+}
+
+export function validateArgs(args: any[], RemoteMethod: any) {
+  let validatesMeta = Reflect.getOwnMetadata(validateMetadataKey, RemoteMethod, 'onRemote') || [];
+  let validations = validatesMeta.map((idx: number) => getValid(args[idx]));
+  return Promise.all(validations);
+}
+
+function getValid(mdl: any) {
+  return new Promise((res: Function, rej: Function) => {
+    mdl.isValid((isValid: boolean) => {
+      if (isValid) {
+        return res();
+      } else {
+        return rej(new ValidationError(mdl));
+      }
+    });
+  });
 }
